Move some logical plan types to expr crate
